requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.4
django-cors-headers==4.7.0
django-redis==5.4.0
django-prometheus==2.4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.27.0
drf-spectacular-sidecar==2025.8.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
packaging==25.0
Pillow==10.4.0
pluggy==1.6.0
prometheus_client==0.22.1
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-django==4.11.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
sqlparse==0.5.3
stripe==12.1.0
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
\n+# Developer typing tools and stubs
pyright
django-stubs
djangorestframework-stubs
